feat: added caching to arvo client when producing messages
